feat(test): refacto et suppression intentions
