+++ pages/index.js
@@ -6,7 +6,7 @@ import {
   STORAGE_LABELS,
   STORAGE_LOCALE,
   STORAGE_SOURCE,
-  STORAGE_TEST_INTENTIONS,
+  STORAGE_TEST_ABC,
 } from "../src/constants/constants"
 import { CATEG, EVENT_CLICK, trackerClick } from "../src/utils/tracker.utils"
 import { useRouter } from "next/router"
@@ -65,7 +65,7 @@ export default function Home() {
   const startSurvey = () => {
     trackerClick(CATEG.home, EVENT_CLICK, `Commencer le test - ${source}`)
     localStorage.setItem(STORAGE_SOURCE, source)
-    localStorage.removeItem(STORAGE_TEST_INTENTIONS)
+    localStorage.removeItem(STORAGE_TEST_ABC)
 
     goToBeforeSurvey()
   }
